+++ systemtest/src/main/java/io/strimzi/systemtest/kafkaClients/api/ClientArgumentMap.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.systemtest.clients.api;
+package io.strimzi.systemtest.kafkaClients.api;
 
 import java.util.ArrayList;
 import java.util.HashMap;
