+++ systemtest/src/main/java/io/strimzi/systemtest/kafkaClients/lib/KafkaClient.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.systemtest.clients.lib;
+package io.strimzi.systemtest.kafkaClients.lib;
 
 import io.vertx.core.Vertx;
 import org.apache.logging.log4j.LogManager;
