+++ systemtest/src/main/java/io/strimzi/systemtest/resources/crd/KafkaResource.java
@@ -24,6 +24,8 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import io.strimzi.systemtest.resources.ResourceManager;
 
+import java.util.function.Consumer;
+
 public class KafkaResource {
     private static final Logger LOGGER = LogManager.getLogger(KafkaResource.class);
 
@@ -186,6 +188,10 @@ public class KafkaResource {
 
     private static Kafka deleteLater(Kafka kafka) {
         return ResourceManager.deleteLater(kafkaClient(), kafka);
+    }
+
+    public static void replaceKafkaResource(String resourceName, Consumer<Kafka> editor) {
+        ResourceManager.replaceCrdResource(Kafka.class, KafkaList.class, DoneableKafka.class, resourceName, editor);
     }
 
 }
