+++ systemtest/src/main/java/io/strimzi/systemtest/resources/crd/KafkaUserResource.java
@@ -17,6 +17,8 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import io.strimzi.systemtest.resources.ResourceManager;
 
+import java.util.function.Consumer;
+
 public class KafkaUserResource {
     private static final Logger LOGGER = LogManager.getLogger(KafkaUserResource.class);
 
@@ -88,5 +90,9 @@ public class KafkaUserResource {
                     .endQuotas()
                 .endSpec()
                 .build());
+    }
+
+    public static void replaceUserResource(String resourceName, Consumer<KafkaUser> editor) {
+        ResourceManager.replaceCrdResource(KafkaUser.class, KafkaUserList.class, DoneableKafkaUser.class, resourceName, editor);
     }
 }
