+++ systemtest/src/test/java/io/strimzi/systemtest/bridge/HttpBridgeST.java
@@ -173,7 +173,7 @@ class HttpBridgeST extends HttpBridgeBaseST {
         StUtils.checkCOlogForUsedVariable(usedVariable);
 
         LOGGER.info("Updating values in Bridge container");
-        replaceBridgeResource(bridgeName, kb -> {
+        KafkaBridgeResource.replaceBridgeResource(bridgeName, kb -> {
             kb.getSpec().getTemplate().getBridgeContainer().setEnv(StUtils.createContainerEnvVarsFromMap(envVarUpdated));
             kb.getSpec().getProducer().setConfig(producerConfig);
             kb.getSpec().getConsumer().setConfig(consumerConfig);
